backend/requirements.txt
fastapi
uvicorn
pypdf
sentence-transformers
chromadb
python-dotenv
requests
torch
ollama
pyjwt
